Refactor animations
